Cleanup tasks
1) remove synchronize

This removes the need to have rsync and all the rsync version
issues.

2) Remove awk
A task to check the existence of a line in file could have been
done with grep (but was problematic on openWRT), so it was done
with awk.
Yet, lineinfile can do the same if we use the check_mode, which
wasn't available when this role was created...

3) Consistent task names
All the tasks were made consistent in their naming,
starting with a capital letter.