Fix openwrt build
On openwrt, the command tincd -K isn't available.
In the past (before the last refactor removing fetch), it wasn't
a problem, as the keys could be generated by the admin on first use,
and then sent to the rest of the cluster from the fetched folder.

The last refactor broke this, and so we need to properly generate
the files and then transfer them

This should fix it.